Fix critical scope error
Chain expressions do modify the current scope, this is desired so we can
run the proper verifications (for example that a class has an specific
property or method). But there is a problem with this aproach when we have
to run verification on arrays or function arguments because the current
scope wasn't being restored to the enclosing one.